Remake ComicDAO into ComicsService and make Comic store multiple authors
ComicDAO was too big and unnecessarily complicated when AuthorDAO and
ComicDAO can be simply used by a ComicsService to get their respective
data and just make a Comic object out of it.
Moreover there was no need to provide a separate DAO just to access the
association table between comic_data and authors.