Add next and previous link to the page displaying a comic page
All links leading to a comic page had to be modified so chapter index
starts from zero. It was previously decremented on the back end to make
the chapter numbers not start from zero but when next and previous link
were added, this would cause the chapter number go lower when called for
the same chapter id. Also Making the frontend increment the value before
sending would unnecessarily couple it with the backend.